+++ src/channels/components/ChannelForm/index.ts
@@ -1,0 +1,2 @@
+export * from "./ChannelForm";
+export { default } from "./ChannelForm";
