+++ src/channels/components/ChannelStatus/index.ts
@@ -1,0 +1,2 @@
+export * from "./ChannelStatus";
+export { default } from "./ChannelStatus";
