+++ src/channels/views/ChannelsList/ChannelsList.tsx
@@ -6,7 +6,6 @@ import ChannelsListPage from "../../pages/ChannelsListPage";
 import { useChannelsList } from "../../queries";
 import {
   channelAddUrl,
-  channelsListUrl,
   ChannelsListUrlQueryParams,
   channelUrl
 } from "../../urls";
@@ -15,7 +14,7 @@ interface ChannelsListProps {
   params: ChannelsListUrlQueryParams;
 }
 
-export const ChannelsList: React.FC<ChannelsListProps> = ({ params }) => {
+export const ChannelsList: React.FC<ChannelsListProps> = () => {
   const navigate = useNavigator();
   const { data, loading } = useChannelsList({ displayLoader: true });
 
@@ -23,25 +22,14 @@ export const ChannelsList: React.FC<ChannelsListProps> = ({ params }) => {
 
   const onRemove = () => null;
 
-  const handleSearchChange = (query: string) => {
-    navigate(
-      channelsListUrl({
-        ...params,
-        query
-      })
-    );
-  };
-
   return (
     <ChannelsListPage
       channelsList={data?.channels}
-      initialSearch={params.query || ""}
       disabled={loading}
       navigateToChannelCreate={navigateToChannelCreate}
       onBack={() => navigate(configurationMenuUrl)}
       onRowClick={id => () => navigate(channelUrl(id))}
       onRemove={onRemove}
-      onSearchChange={handleSearchChange}
     />
   );
 };
